RESTEASY-800: Renamed arquillian/RESTEASY-802-AS71 to RESTEASY-800-AS71.
