+++ arquillian/RESTEASY-800-AS71/src/main/java/org/jboss/resteasy/resteasy800/TestApplication.java
@@ -1,4 +1,4 @@
-package org.jboss.resteasy.resteasy802;
+package org.jboss.resteasy.resteasy800;
 
 import java.util.HashSet;
 import java.util.Set;
