+++ arquillian/RESTEASY-800-AS71/src/main/java/org/jboss/resteasy/resteasy800/TestResource.java
@@ -1,4 +1,4 @@
-package org.jboss.resteasy.resteasy802;
+package org.jboss.resteasy.resteasy800;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
