+++ arquillian/RESTEASY-800-AS71/src/test/resources/arquillian.xml
